Address type resolution in publisher not server
Currently we're doing address type resolution in the server, which means
if we were to choose 'public' as the address to bind to, the server
would find and use a public address, but the publisher would report
'http://public:....'.

This PR moves address type resolution to the publisher itself, so that
it can compose a URL with the correct IP, and the server is given the
correct IP at the beginning.

Fixes #3554
